Fixed delete for TemporaryApplication. All find and delete are now working. Refactored Application to ApplicationDeployment
