+++ src/main/kotlin/no/skatteetaten/aurora/gorg/controller/ApplicationController.kt
@@ -1,7 +1,7 @@
 package no.skatteetaten.aurora.gorg.controller
 
-import no.skatteetaten.aurora.gorg.service.ApplicationService
-import no.skatteetaten.aurora.gorg.service.TemporaryApplication
+import no.skatteetaten.aurora.gorg.service.ApplicationDeploymentService
+import no.skatteetaten.aurora.gorg.service.TemporaryApplicationDeployment
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
@@ -10,16 +10,16 @@ import java.time.Instant
 
 @RestController
 @RequestMapping("/api/apps")
-class ApplicationController(val applicationService: ApplicationService) {
+class ApplicationController(val applicationDeploymentService: ApplicationDeploymentService) {
 
     @DeleteMapping
     fun deleteApplications() {
-        applicationService.findTemporaryApplications(Instant.now())
+        applicationDeploymentService.findTemporaryApplicationDeployments(Instant.now())
                 .filter { it.ttl.isNegative }
-                .forEach { applicationService.deleteApplication(it) }
+                .forEach { applicationDeploymentService.deleteApplicationDeployment(it) }
     }
 
     @GetMapping
-    fun list(): List<TemporaryApplication> =
-        applicationService.findTemporaryApplications(Instant.now())
+    fun list(): List<TemporaryApplicationDeployment> =
+        applicationDeploymentService.findTemporaryApplicationDeployments(Instant.now())
 }
