+++ src/main/kotlin/no/skatteetaten/aurora/gorg/extensions/DefaultOpenShiftClient.kt
@@ -7,25 +7,25 @@ import io.fabric8.kubernetes.api.model.ObjectMeta
 import io.fabric8.kubernetes.client.KubernetesClientException
 import io.fabric8.openshift.client.DefaultOpenShiftClient
 import okhttp3.Request
+import okhttp3.Response
 
-/*
-fun DefaultOpenShiftClient.deleteApplicationDeployment(namespace:String, name:String): ApplicationDeployment {
+fun DefaultOpenShiftClient.deleteApplicationDeployment(namespace:String, name:String): Response {
     val url =
         this.openshiftUrl.toURI().resolve("/apis/skatteetaten.no/v1/namespaces/$namespace/applicationdeployments/$name")
     return try {
-        val request = Request.Builder().url(url.toString()).build() //set verb
+        val request = Request.Builder().url(url.toString()).delete().build()
         val response = this.httpClient.newCall(request).execute()
-        jacksonObjectMapper().readValue(response.body()?.bytes(), ApplicationDeployment::class.java)
-            ?: throw KubernetesClientException("Error occurred while fetching temporary application deployments")
+            response
     } catch (e: Exception) {
         throw KubernetesClientException("Error occurred while fetching temporary application deployments", e)
     }
 }
-*/
+
 
 fun DefaultOpenShiftClient.applicationDeploymentsTemporary(): List<ApplicationDeployment> {
     val url =
         this.openshiftUrl.toURI().resolve("/apis/skatteetaten.no/v1/applicationdeployments?labelSelector=ttl")
+
     return try {
         val request = Request.Builder().url(url.toString()).build()
         val response = this.httpClient.newCall(request).execute()
